+++ Source/EventFlow/ReadStores/IReadModelPopulator.cs
@@ -21,6 +21,7 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -31,13 +32,15 @@ namespace EventFlow.ReadStores
         Task PurgeAsync<TReadModel>(CancellationToken cancellationToken)
             where TReadModel : class, IReadModel, new();
 
-        void Purge<TReadModel>(CancellationToken cancellationToken)
-            where TReadModel : class, IReadModel, new();
-
         Task PopulateAsync<TReadModel>(CancellationToken cancellationToken)
             where TReadModel : class, IReadModel, new();
 
-        void Populate<TReadModel>(CancellationToken cancellationToken)
-            where TReadModel : class, IReadModel, new();
+        Task PopulateAsync(
+            Type readModelType,
+            CancellationToken cancellationToken);
+
+        Task PurgeAsync(
+            Type readModelType,
+            CancellationToken cancellationToken);
     }
 }